requirements.txt
flask==3.0.3
torch==2.3.0
torchvision==0.18.0
Pillow==10.3.0
numpy==1.26.4
matplotlib==3.9.0
seaborn==0.13.2
scikit-learn==1.5.0
tqdm==4.66.4
werkzeug==3.0.3